Dodanie flag. Nie dziala. Nie wiem dlaczego.

+++ fault.h
@@ -25,7 +25,7 @@ public:
 	bool Check(uint8_t fault, uint8_t address);
 	bool CheckAll(uint8_t address);
 private:
-	uint32_t global_faults;
+	uint8_t global_faults;
 	uint8_t  doors_faults[MACHINE_MAX_NUMBER_OF_DOORS];
 };
